+++ packages/firestore/src/core/firestore_client.ts
@@ -22,7 +22,6 @@ import { GarbageCollectionScheduler, Persistence } from '../local/persistence';
 import { Document, NoDocument } from '../model/document';
 import { DocumentKey } from '../model/document_key';
 import { Mutation } from '../model/mutation';
-import { loadConnection, newSerializer } from '../platform/platform';
 import { newDatastore } from '../remote/datastore';
 import { RemoteStore } from '../remote/remote_store';
 import { AsyncQueue, wrapInUserErrorIfRecoverable } from '../util/async_queue';
@@ -48,6 +47,8 @@ import {
   ComponentProvider,
   MemoryComponentProvider
 } from './component_provider';
+import { newConnection } from '../platform/connection';
+import { newSerializer } from '../platform/serializer';
 
 const LOG_TAG = 'FirestoreClient';
 const MAX_CONCURRENT_LIMBO_RESOLUTIONS = 100;
@@ -234,7 +235,7 @@ export class FirestoreClient {
       // Datastore (without duplicating the initializing logic once per
       // provider).
 
-      const connection = await loadConnection(this.databaseInfo);
+      const connection = await newConnection(this.databaseInfo);
       const serializer = newSerializer(this.databaseInfo.databaseId);
       const datastore = newDatastore(connection, this.credentials, serializer);
