+++ packages/firestore/src/platform/browser/format_json.ts
