+++ packages/firestore/src/platform/node/format_json.ts
