+++ packages/firestore/src/platform/rn/dom.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2017 Google LLC
+ * Copyright 2020 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,12 +15,4 @@
  * limitations under the License.
  */
 
-import { isNode } from '@firebase/util';
-import { hardAssert } from '../util/assert';
-
-hardAssert(
-  isNode(),
-  'The generic Platform implementation should only run under ts-node.'
-);
-
-export { randomBytes } from './node/random_bytes';
+export { getWindow, getDocument } from '../browser/dom';
