--- packages/firestore/src/platform_node/node_platform.ts
@@ -1,85 +1,0 @@
-/**
- * @license
- * Copyright 2017 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { DatabaseId, DatabaseInfo } from '../core/database_info';
-import { Connection } from '../remote/connection';
-import { JsonProtoSerializer } from '../remote/serializer';
-import { Code, FirestoreError } from '../util/error';
-import { ConnectivityMonitor } from '../remote/connectivity_monitor';
-import { NoopConnectivityMonitor } from '../remote/connectivity_monitor_noop';
-
-import { GrpcConnection } from './grpc_connection';
-import { loadProtos } from './load_protos';
-
-// The exports in this class must match the exports in '../platform/platform' as
-// they are bundled with the browser build during the Rollup build.
-
-/** Loads the GRPC stack */
-export function loadConnection(
-  databaseInfo: DatabaseInfo
-): Promise<Connection> {
-  const protos = loadProtos();
-  return Promise.resolve(new GrpcConnection(protos, databaseInfo));
-}
-
-/** Return the Platform-specific connectivity monitor. */
-export function newConnectivityMonitor(): ConnectivityMonitor {
-  return new NoopConnectivityMonitor();
-}
-
-/** Return the Platform-specific serializer monitor. */
-export function newSerializer(databaseId: DatabaseId): JsonProtoSerializer {
-  return new JsonProtoSerializer(databaseId, { useProto3Json: false });
-}
-
-/** Converts a Base64 encoded string to a binary string. */
-export function decodeBase64(encoded: string): string {
-  // Node actually doesn't validate base64 strings.
-  // A quick sanity check that is not a fool-proof validation
-  if (/[^-A-Za-z0-9+/=]/.test(encoded)) {
-    throw new FirestoreError(
-      Code.INVALID_ARGUMENT,
-      'Not a valid Base64 string: ' + encoded
-    );
-  }
-  return new Buffer(encoded, 'base64').toString('binary');
-}
-
-/** Converts a binary string to a Base64 encoded string. */
-export function encodeBase64(raw: string): string {
-  return new Buffer(raw, 'binary').toString('base64');
-}
-
-/** The Platform's 'window' implementation or null if not available. */
-export function getWindow(): Window | null {
-  if (process.env.USE_MOCK_PERSISTENCE === 'YES') {
-    // eslint-disable-next-line no-restricted-globals
-    return window;
-  }
-
-  return null;
-}
-
-/** The Platform's 'document' implementation or null if not available. */
-export function getDocument(): Document | null {
-  return null;
-}
-
-/** True if and only if the Base64 conversion functions are available. */
-export function isBase64Available(): boolean {
-  return true;
-}
